feat: add redux support across bridge
The default `index.ts` that gets imported is the entry point for the Model
thread, and it default exports the bridge instance. Similarly, `/src/view.ts` is
the entry point for the View thread, and it default exports that side of the
bridge.

Both bridges need to have their I/O configured, and can be extended if needed.